feat/setup: add the ESLint rules & change name file tailwind.config & moves types to api.d.ts
