+++ .eslintrc.json
@@ -1,3 +1,13 @@
 {
-  "extends": ["next/core-web-vitals", "next/typescript"]
+  "extends": ["next/core-web-vitals", "next/typescript", "prettier"],
+  "rules": {
+    "no-multiple-empty-lines": ["error", { "max": 1, "maxEOF": 1 }],
+    "semi": ["error", "always"],
+    "quotes": ["error", "double"],
+    "eqeqeq": ["error", "always"],
+    "no-unused-vars": "warn",
+    "no-console": "warn",
+    "import/order": ["error", { "groups": ["builtin", "external", "internal"] }],
+    "@typescript-eslint/no-explicit-any": "warn"
+  }
 }
