+++ src/components/ui/label.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 import * as LabelPrimitive from "@radix-ui/react-label";
 import { cva, type VariantProps } from "class-variance-authority";
 
-import { cn } from "@/lib/utils/tailwind-merge";
+import { cn } from "@/lib/utils/cn";
 
 const labelVariants = cva(
   "text-sm font-bold leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70",
