+++ src/components/ui/sonner.tsx
@@ -1,6 +1,8 @@
-import { cn } from "@/lib/utils/tailwind-merge";
+"use client";
+
+import { cn } from "@/lib/utils/cn";
 import { useTheme } from "next-themes";
-import { Toaster as Sonner } from "sonner"; // من ShadCN برضو
+import { Toaster as Sonner } from "sonner";
 
 type ToasterProps = React.ComponentProps<typeof Sonner>;
