--- src/lib/types/next-auth.d.ts
@@ -1,14 +1,0 @@
-declare type DatabaseProperties = {
-  _id: string;
-  createdAt: string;
-};
-
-declare type SuccessfulResponse<T> = {
-  message: "success";
-} & T;
-
-declare type ErrorResponse = {
-  error: string;
-};
-
-declare type APIResponse<T> = SuccessfulResponse<T> | ErrorResponse;
